# docs/openapi/compute-service.yaml
...
components:
  schemas:
...
    Corev1Toleration:
      properties:
        effect:
          type: string
        key:
          type: string
        operator:
          type: string
        tolerationSeconds:
          format: int64
          nullable: true
          type: integer
        value:
          type: string
      type: object
...
    Problem:
      properties:
        code:
          description: Discrete business error class.
          enum:
          - validation_failed
          - not_found
          - conflict
          - precondition_failed
          - unauthorized
          - forbidden
          - service_unavailable
          - internal_error
          - quota_exceeded
          type: string
        detail:
          type: string
        details:
          additionalProperties: {}
          type: object
        instance:
          type: string
        status:
          type: integer
        title:
          type: string
        type:
          type: string
      required:
      - code
      - status
      - title
      - type
      type: object
    ResourcepoolCreateInput:
      properties:
        description:
          type: string
        metadata:
          additionalProperties: {}
          type: object
        name:
          maxLength: 40
          minLength: 3
          pattern: ^[a-z0-9](?:[a-z0-9-]{1,38}[a-z0-9])?$
          type: string
        nodeSelector:
          additionalProperties:
            type: string
          type: object
        tolerations:
          items:
            $ref: '#/components/schemas/Corev1Toleration'
          type: array
      required:
      - name
      type: object
    ResourcepoolList:
      properties:
        items:
          items:
            $ref: '#/components/schemas/ResourcepoolView'
          type: array
        total:
          format: int64
          type: integer
      required:
      - items
      - total
      type: object
    ResourcepoolUpdateInput:
      properties:
        description:
          nullable: true
          type: string
        metadata:
          additionalProperties: {}
          nullable: true
          type: object
        nodeSelector:
          additionalProperties:
            type: string
          nullable: true
          type: object
        tolerations:
          items:
            $ref: '#/components/schemas/Corev1Toleration'
          nullable: true
          type: array
      type: object
    ResourcepoolView:
      properties:
        createdAt:
          format: date-time
          type: string
        description:
          type: string
        id:
          format: uuid
          type: string
        metadata:
          additionalProperties: {}
          type: object
        name:
          type: string
        nodeSelector:
          additionalProperties:
            type: string
          type: object
        tolerations:
          items:
            $ref: '#/components/schemas/Corev1Toleration'
          type: array
        updatedAt:
          format: date-time
          type: string
      required:
      - createdAt
      - id
      - name
      - updatedAt
      type: object
...
info:
  description: REST API for resource pools, resource units, jobs, and services. Partitioned
    by Kubernetes namespace. RFC7807 Problem responses on errors.
  title: AxisML Compute Service API
  version: 0.1.0
openapi: 3.0.3
paths:
...
  /api/v1/resource-pools:
    get:
      operationId: listResourcePools
      parameters:
      - description: Page size (1–200, default 50).
        in: query
        name: limit
        schema:
          format: int32
          type: integer
      - description: Number of items to skip.
        in: query
        name: offset
        schema:
          format: int32
          type: integer
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcepoolList'
          description: Page.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: List resource pools
      tags:
      - resource-pools
    post:
      operationId: createResourcePool
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourcepoolCreateInput'
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcepoolView'
          description: Created.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Create a resource pool
      tags:
      - resource-pools
  /api/v1/resource-pools/{pool}:
    delete:
      operationId: deleteResourcePool
      parameters:
      - description: Resource pool name.
        in: path
        name: pool
        required: true
        schema:
          type: string
      responses:
        "204":
          description: Deleted (idempotent).
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Delete resource pool
      tags:
      - resource-pools
    get:
      operationId: getResourcePool
      parameters:
      - description: Resource pool name.
        in: path
        name: pool
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcepoolView'
          description: Resource pool.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Get resource pool
      tags:
      - resource-pools
    patch:
      operationId: updateResourcePool
      parameters:
      - description: Resource pool name.
        in: path
        name: pool
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResourcepoolUpdateInput'
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResourcepoolView'
          description: Updated.
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Validation error.
        "401":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unauthorized.
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Forbidden.
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Not found.
        "409":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Conflict.
        "412":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Precondition failed.
        "422":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Quota exceeded.
        "503":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Service unavailable.
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Problem'
          description: Unexpected error.
      summary: Patch resource pool
      tags:
      - resource-pools
...
servers:
- description: Same-origin
  url: /
...